AJAX like button and basic user edit form

+++ src/user_profile/models.py
@@ -7,6 +7,8 @@ from datetime import timedelta
 from django.dispatch import receiver
 from django.db import models
 from django.contrib.auth.models import AbstractUser
+from django.shortcuts import reverse
+from django.templatetags.static import static
 
 # from moments_feed.models import Like
 from utils.img_tools import compress_image
@@ -25,6 +27,15 @@ class User(AbstractUser):
     #     likes = Like.objects.filter(created_date__range=[start_date, end_date])
     #     return likes
 
+    def get_avatar_url(self):
+        if self.avatar:
+            return self.avatar.url
+        else:
+            return static('default-avatar.jpg')
+
+    def get_absolute_url(self):
+        return reverse('profile', args=(self.username, ))
+
     def save(self, *args, **kwargs):
         super().save(*args, **kwargs)
         if self.avatar:
@@ -33,9 +44,9 @@ class User(AbstractUser):
 
 @receiver(models.signals.post_delete, sender=User)
 def auto_delete_image_on_delete(sender, instance, **kwargs):
-    if instance.image:
-        if os.path.isfile(instance.image.path):
-            os.remove(instance.image.path)
+    if instance.avatar:
+        if os.path.isfile(instance.avatar.path):
+            os.remove(instance.avatar.path)
 
 
 @receiver(models.signals.pre_save, sender=User)
@@ -48,6 +59,6 @@ def auto_delete_file_on_change(sender, instance, **kwargs):
     except User.DoesNotExist:
         return False
 
-    if instance.avatar != old_avatar:
+    if instance.avatar and old_avatar and instance.avatar != old_avatar:
         if os.path.isfile(old_avatar.path):
             os.remove(old_avatar.path)
